HIPANGULAR-23 First version of bundeling (from 400 to 70 request)
