+++ app/authentication/authentication.routing.ts
@@ -1,0 +1,18 @@
+import { ModuleWithProviders } from '@angular/core';
+import { Routes, RouterModule }   from '@angular/router';
+
+import { LoginComponent } from './login/login.component';
+import { SignupComponent } from './signup/signup.component';
+
+const authRoutes: Routes = [
+  {
+    path: 'login',
+    component: LoginComponent
+  },
+  {
+    path: 'signup',
+    component: SignupComponent
+  },
+];
+
+export const authRouting: ModuleWithProviders = RouterModule.forChild(authRoutes);
