+++ app/footer/contact/contact.component.ts
@@ -2,7 +2,7 @@ import { Component } from '@angular/core';
 
 @Component({
   selector: 'hip-contact',
-  templateUrl: './app/contact/contact.component.html',
+  templateUrl: './app/footer/contact/contact.component.html',
 })
 export class ContactComponent {
 }
