--- app/footer/footer.component.css
@@ -1,19 +1,0 @@
-footer {
-  background: #00205B;
-  height: 70px;
-  padding: 15px;
-  width: 100%;
-
-}
-
-.legal {
-  max-width: 1024px;
-  text-align: center;
-  color: #FFFFFF;
-  margin: 0 auto;
-}
-
-.legal a {
-  color: #FFFFFF;
-  text-decoration: none;
-}
