--- app/footer/footer.component.html
@@ -1,8 +1,0 @@
-<footer>
-  <div class="legal">
-      <a [routerLink]="[ '/contact' ]">{{ 'contact' | translate }}</a>
-      | <a [routerLink]="[ '/help' ]">{{ 'help' | translate }}</a>
-      | <a [routerLink]="[ '/legal-notice' ]">{{ 'legal notice' | translate }}</a>
-    <br/><span>{{hipCopyright}}</span>
-  </div>
-</footer>
