--- app/footer/footer.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-    selector: 'hip-footer',
-    templateUrl: './app/footer/footer.component.html',
-    styleUrls: ['./app/footer/footer.component.css']
-})
-export class FooterComponent {
-    hipCopyright = 'HiP CMS';
-}
