+++ app/footer/footer.module.ts
@@ -1,0 +1,28 @@
+import { NgModule } from '@angular/core';
+import { BrowserModule } from '@angular/platform-browser';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { MaterialModule } from '@angular/material';
+
+import { footerRouting } from './footer.routing';
+import { ContactComponent } from './contact/contact.component';
+import { LegalNoticeComponent } from './legal-notice/legal-notice.component';
+import { HelpComponent } from './help/help.component';
+import { TranslateModule } from 'ng2-translate';
+
+@NgModule({
+  imports: [
+    BrowserModule,
+    FormsModule,
+    ReactiveFormsModule,
+    MaterialModule,
+    TranslateModule,
+    footerRouting,
+  ],
+  declarations: [
+    ContactComponent,
+    LegalNoticeComponent,
+    HelpComponent
+  ]
+})
+export class FooterModule {
+}
