+++ app/footer/footer.routing.ts
@@ -1,0 +1,24 @@
+import { ModuleWithProviders } from '@angular/core';
+import { Routes, RouterModule }   from '@angular/router';
+
+import { ContactComponent } from './contact/contact.component';
+import { LegalNoticeComponent } from './legal-notice/legal-notice.component';
+import { HelpComponent } from './help/help.component';
+
+
+const footerRoutes: Routes = [
+  {
+    path: 'contact',
+    component: ContactComponent
+  },
+  {
+    path: 'legal-notice',
+    component: LegalNoticeComponent
+  },
+  {
+    path: 'help',
+    component: HelpComponent
+  }
+];
+
+export const footerRouting: ModuleWithProviders = RouterModule.forChild(footerRoutes);
