+++ app/footer/help/index.ts
