+++ app/footer/index.ts
@@ -1,1 +1,1 @@
-export * from './footer.component';
+export * from './';
