+++ app/footer/legal-notice/index.ts
