+++ app/footer/legal-notice/legal-notice.component.html
