+++ app/footer/legal-notice/legal-notice.component.ts
@@ -2,7 +2,7 @@ import { Component } from '@angular/core';
 
 @Component({
     selector: 'hip-legal-notice',
-    templateUrl: './app/legal-notice/legal-notice.component.html'
+  templateUrl: './app/footer/legal-notice/legal-notice.component.html'
 })
 export class LegalNoticeComponent {
 }
