--- app/sidenav/sidenav.component.html
@@ -1,22 +1,0 @@
-<hip-toolbar [start]="start" [sidenavOpened]="opened"></hip-toolbar>
-<md-sidenav-layout flex>
-  <md-sidenav #start [mode]="mode" [opened]="opened">
-    <md-nav-list>
-      <md-list-item *ngFor="let element of navigation">
-        <h2 md-line>
-          <a [routerLink]="[ element.link ]"
-             (click)="start.toggle()" *ngIf="!opened">{{element.name | lowercase | translate}}</a>
-          <a [routerLink]="[ element.link ]"
-             *ngIf="opened">{{element.name | lowercase | translate}}</a>
-        </h2>
-      </md-list-item>
-    </md-nav-list>
-    <br>
-  </md-sidenav>
-  <div class="wrapper">
-    <toaster-container></toaster-container>
-    <router-outlet></router-outlet>
-    <div class="push"></div>
-  </div>
-  <hip-footer></hip-footer>
-</md-sidenav-layout>
