+++ app/topics/topics.module.ts
@@ -27,7 +27,7 @@ import { AddNewSubtopicComponent } from './topic-management/add-new-subtopic/add
 import { TagInputComponent } from '../shared/taginput/taginput.component';
 import { AllTopicsComponent } from './all-topics/all-topics.component';
 import { TreeView } from './shared/treeview-node/treeview-node.component';
-import { TopicsFilterPipe } from '../topics/pipes/topic-filter.pipe';
+import { TopicsFilterPipe } from './/pipes/topic-filter.pipe';
 import { AddExistingSubtopicComponent } from './topic-management/add-existing-subtopic/add-existing-subtopic.component';
 import { ContentComponent } from './content/content.component';
