+++ index.html
@@ -22,15 +22,10 @@
   <script src="node_modules/reflect-metadata/Reflect.js"></script>
   <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.11.3/jquery.min.js"></script>
   <script src="assets/bootstrap/js/bootstrap.min.js"></script>
+  <script src="/assets/only-office/api.js"></script>
 
-  <!-- ng2-pagination for pagination -->
-  <script src="node_modules/ng2-pagination/dist/ng2-pagination-bundle.js"></script>
+  <!--<script src="systemjs.config.js"></script>-->
+  <script src="app/bundle.js"></script>
 
-  <script src="systemjs.config.js"></script>
-  <script>
-    System.import('app').catch(function (err) {
-      console.error(err);
-    });
-  </script>
   </body>
 </html>
