+++ package.json
@@ -20,7 +20,9 @@
     "doc": "typedoc --out \"docs/\" app/",
     "clean": "rimraf hip-theme.css typings node_modules || exit 0",
     "clean-install": "npm run clean && npm install",
-    "sass": "node-sass hip-theme.scss hip-theme.css"
+    "sass": "node-sass hip-theme.scss hip-theme.css",
+    "build_prod": "tsc && browserify --ignore-missing -s main app/main.js > app/bundle.js",
+    "minify": "uglifyjs app/bundle.js --screw-ie8 --compress --mangle --output app/bundle.min.js"
   },
   "keywords": [],
   "author": "",
@@ -51,6 +53,7 @@
   "devDependencies": {
     "@angular/compiler-cli": "2.2.3",
     "@angular/platform-server": "2.2.3",
+    "sass-loader": "^4.0.2",
     "canonical-path": "0.0.2",
     "codelyzer": "1.0.0-beta.4",
     "concurrently": "3.1.0",
@@ -65,6 +68,7 @@
     "lite-server": "2.2.0",
     "lodash": "4.17.2",
     "node-sass": "3.13.0",
+    "postcss-loader": "^1.1.1",
     "protractor": "4.0.11",
     "rimraf": "2.5.2",
     "tslint": "3.15.1",
